fix: green the full gate — 5 targets broken under --all-targets --all-features
Follow-on from c284dbb2. Fixing Issue 700 R1 exposed the generalisation: a
non-default feature's code compiles to nothing, so no gate the repo actually
ran was compiling it. Running the strongest form found 5 broken targets that
every existing green gate had been silently skipping:

  cargo clippy --workspace --all-targets --all-features --keep-going

`--keep-going` mattered — without it the run stopped after 2 of the 5.

Three of these are the trio proven pre-existing at HEAD 26d055c6 back in
96543f4f and left unfixed; two are newly surfaced. All five are the same
meta-class as R1 — drift in code no gate compiles — and two are literally the
same heal-sweep escape classes:

- tests/verify_tiktoken_fix.rs:68 (E0425) — `manual_let_else` rename miss,
  identical to the R1 defect: binding became `d`, use site still said `data`.
- tests/nf_flow_goat.rs:449 (E0631/E0599) — `redundant_closure` applied across
  a deref coercion. `.map(categorical_entropy)` cannot work: the iterator
  yields `&Vec<f32>` and the fn takes `&[f32]`; coercion is available to a call
  expression, not to a fn item passed as `FnMut`. Fixed with an explicit
  `.as_slice()` rather than `|p| categorical_entropy(p)`, which clippy would
  re-flag — closing the loop instead of re-entering it.
- crates/katgpt-core/tests/bench_688_certified_frontier_goat.rs:199 (E0308) —
  `Vec::with_capacity(SEEDS)`, SEEDS is u64; the file already casts at its two
  other use sites.
- examples/bomber_02_tui.rs:139 (E0308) — same shape, TICK_LIMIT is u32.
- crates/katgpt-core/src/anchored_reach.rs:215 — deny-level `approx_constant`
  on SWEEP's 3.14159. That value is an arbitrary non-power-of-two decimal
  chosen as a bit-identity hazard, NOT an approximation of pi, so this is a
  module-level allow with the rationale recorded. Substituting `consts::PI`
  would silently change a pinned test vector.

AGENTS.md gains a "full gate" subsection under Build Commands: every command
previously listed there is narrow on at least one of three independent axes
(check-vs-clippy, default-vs-all-features, -p-vs---workspace), which is why
this gate could stay red from 2cb97410 until now with all of them green.

Validation:
- `cargo clippy --workspace --all-targets --all-features --keep-going` — 0
  errors, 0 broken targets (was 5). First green run of this command.
- Repaired targets pass, not merely compile: bench_688 3/0, anchored_reach 7/0,
  verify_tiktoken_fix 2/0 (fixture present — not a vacuous early-return),
  nf_flow_goat 7/0, bomber_02_tui builds.
- `cargo test --workspace --lib` — 32 binaries, 0 failed.
- Diff is 20 insertions / 5 deletions across 5 files. An earlier
  `cargo fmt -p katgpt-rs -- <files>` ignored its file arguments and
  reformatted 89 files; that was reverted in full and the edits re-applied by
  hand, so no unrelated formatting rides along.

+++ crates/katgpt-core/tests/bench_688_certified_frontier_goat.rs
@@ -196,7 +196,7 @@ fn t3_4_report_the_floor_adjacency_only_expansion() {
     );
 
     let mut prim_wins = 0usize;
-    let mut ratios = Vec::with_capacity(SEEDS);
+    let mut ratios = Vec::with_capacity(SEEDS as usize);
     let (mut prim_viol, mut floor_viol) = (0usize, 0usize);
     let (mut prim_cert, mut floor_cert) = (0usize, 0usize);
